refactor: remove sub-packages
