+++ nuv/util.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 //
-package utils
+package main
 
 import (
 	"errors"
